New yaml, controller and service file for fabric admin
