+++ packages/fabric-backend/lib/src/main/resources/api/fabricWorkspaces.yaml
@@ -525,56 +525,6 @@ paths:
         500:
           description: "Internal server error."
 
-  /fabric-workspaces/admin:
-    get:
-      tags:
-        - "fabric-workspaces"
-      summary: "Get all workspaces for Fabric Admin users."
-      description: "Fetches all valid fabric workspace records. This endpoint is available only for users who have the FabricAdmin role."
-      operationId: "getAllForFabricAdmin"
-      parameters:
-        - name: offset
-          in: query
-          description: Page number from which listing of workspaces should start. Offset. Example 2
-          required: false
-          type: integer
-        - name: limit
-          in: query
-          description: Page size to limit the number of workspaces, Example 15
-          required: false
-          type: integer
-        - name: sortBy
-          in: query
-          description: Sort workspaces by a given variable like name or createdOn
-          required: false
-          type: string
-          enum: [name, createdOn]
-        - name: sortOrder
-          in: query
-          description: Sort workspaces based on the given order (asc or desc)
-          required: false
-          type: string
-          enum: [asc, desc]
-      responses:
-        200:
-          description: "Returns list of workspaces for Fabric Admin."
-          schema:
-            $ref: "#/definitions/FabricWorkspacesCollectionVO"
-        204:
-          description: "Fetch complete, no content found."
-        400:
-          description: "Bad request."
-        401:
-          description: "Request does not have sufficient credentials."
-        403:
-          description: "User is not a Fabric Admin."
-        405:
-          description: "Method not allowed."
-        500:
-          description: "Internal server error."
-
-
-
 definitions:
   CapacityVO:
     type: object
