+++ packages/fabric-backend/lib/src/main/resources/api/fabricWorkspacesAdmin.json
@@ -1,0 +1,15 @@
+{
+	"library": "spring-mvc",
+	"dateLibrary": "legacy",
+	"hideGenerationTimestamp": true,
+	"useTags": true,
+	"modelPackage": "com.daimler.data.dto.fabricWorkspaceAdmin",
+	"apiPackage": "com.daimler.data.api.fabricWorkspaceAdmin",
+	"invokerPackage": "com.daimler.data.service.fabricWorkspaceAdmin",
+	"importMappings": {
+		"BindingResult": "org.springframework.validation.BindingResult",
+		"GenericMessage": "com.daimler.data.controller.exceptions.GenericMessage",
+		"MessageDescription": "com.daimler.data.controller.exceptions.MessageDescription",
+		"FabricWorkspacesCollectionVO": "com.daimler.data.dto.fabricWorkspace.FabricWorkspacesCollectionVO"
+	}
+}
