+++ packages/fabric-backend/lib/src/main/resources/api/fabricWorkspacesAdmin.yaml
@@ -1,0 +1,116 @@
+# LICENSE START
+# 
+# MIT License
+# 
+# Copyright (c) 2019 Daimler TSS GmbH
+# 
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+# 
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+# 
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+# 
+# LICENSE END 
+#
+#-----------------------------------------------------------------------------------------------------
+#  **
+#  ** File Name  : fabricWorkspaces.yaml
+#  **
+#  ** Module     : Tools/FabricWorkspaces
+#  **
+
+#  -------------------------------------------------------------------------------------------------------
+#
+#  CHANGE HISTORY
+#
+#  Created By         :  boonbaluri
+#
+#  Bug Number         :
+#  Changed By         :
+#  Change Description :
+#  Change On          :
+#  Search String      :
+#
+swagger: "2.0"
+info:
+  version: "1.0.0"
+  title: "Fabric workspaces for Admin management service"
+host: "localhost:9292"
+basePath: "/api/1.0"
+consumes:
+  - "application/json"
+produces:
+  - "application/json"
+tags:
+  - name: "solutions"
+    description: "This service is used for fabric workspaces maintenance."
+
+schemes:
+  - "https"
+
+paths:
+  # path to create a new solution and fetch all valid solution or by intent
+  #
+  #  supprorted http methods GET, POST and DELETE
+  #
+  # path example ../fabric-workspaces
+  /fabric-workspaces/admin:
+    get:
+      tags:
+        - "fabric-workspaces-admin"
+      summary: "Get all workspaces for Fabric Admin users."
+      description: "Fetches all valid fabric workspace records. This endpoint is available only for users who have the FabricAdmin role."
+      operationId: "getAllForFabricAdmin"
+      parameters:
+        - name: offset
+          in: query
+          description: Page number from which listing of workspaces should start. Offset. Example 2
+          required: false
+          type: integer
+        - name: limit
+          in: query
+          description: Page size to limit the number of workspaces, Example 15
+          required: false
+          type: integer
+        - name: sortBy
+          in: query
+          description: Sort workspaces by a given variable like name or createdOn
+          required: false
+          type: string
+          enum: [name, createdOn]
+        - name: sortOrder
+          in: query
+          description: Sort workspaces based on the given order (asc or desc)
+          required: false
+          type: string
+          enum: [asc, desc]
+      responses:
+        200:
+          description: "Returns list of workspaces for Fabric Admin."
+          schema:
+            $ref: "#/definitions/FabricWorkspacesCollectionVO"
+        204:
+          description: "Fetch complete, no content found."
+        400:
+          description: "Bad request."
+        401:
+          description: "Request does not have sufficient credentials."
+        403:
+          description: "User is not a Fabric Admin."
+        405:
+          description: "Method not allowed."
+        500:
+          description: "Internal server error."
+
